+++ app/assets/js/src/images/index.ts
@@ -1,0 +1,7 @@
+export { getImage } from './getImage';
+export { hasImage } from './hasImage';
+export { saveImage } from './saveImage';
+
+export { createImageUrlPlaceholder } from './createImageUrlPlaceholder';
+export { consumeImageUrlPlaceholder } from './consumeImageUrlPlaceholder';
+export { consumeAllImageUrlPlaceholders } from './consumeAllImageUrlPlaceholders';